refactor the cache policies
